move logic to function
